requirements.txt
fastapi
uvicorn
httpx
pydantic
python-multipart
python-dotenv
pytest
pytest-asyncio
